Add many improvements: enable encryption/decryption and make code platform independent

+++ Code/AutoCyberoamLogin.rb
@@ -1,4 +1,118 @@
-require 'cipher.rb'
+$windows = Gem.win_platform?
+
+Shoes.setup do
+ gem 'nori'
+ gem 'nokogiri'
+ gem 'gibberish' 
+ gem 'libnotify' if $windows
+end
+require 'nori'
+require 'nokogiri'
+require 'libnotify'
+require 'gibberish' if $windows
+
+
+#-----Notification related methods----------#
+    def show_notification(title , body)
+      if not $windows
+        notification = Libnotify.new do |notify|
+          notify.summary    = title
+          notify.body       = body
+          notify.timeout    = 1         
+          notify.urgency    = :critical   
+          notify.append     = true       
+          notify.transient  = false        
+        end
+        notification.show!
+      end
+  end
+#-------------------------------------------#
+
+
+#-----User Information Methods--------------#
+    def cache_details(username , password)   
+       repeating_seconds = get_repeating_seconds
+       file = File.new($filepath ,"w:UTF-8") 
+         file.puts "username::>#{username}"
+         file.puts "password::>#{encrypt(password)}"
+         file.puts "repeating_seconds::>#{repeating_seconds}" if not repeating_seconds.strip.empty?
+       file.close
+    end
+
+    def save_repeating_time(repeating_seconds)
+       username = get_username
+       password = get_password
+       if( (not username.empty?) and (not password.empty?) )
+           file = File.new($filepath , "w:UTF-8") 
+             file.puts "username::>#{username}"
+             file.puts "password::>#{encrypt(password)}"
+             file.puts "repeating_seconds::>#{repeating_seconds}"
+           file.close
+           return true
+        else
+           return false
+        end
+    end
+    def unsave_repeating_time
+       username = get_username
+       password = get_password
+       if( (not username.empty?) and (not password.empty?) )
+           file = File.new($filepath , "w:UTF-8") 
+             file.puts "username::>#{username}"
+             file.puts "password::>#{encrypt(password)}"
+           file.close
+        end
+    end
+    def get_username
+      return "" if not File.exists?($filepath)
+      file = File.new( $filepath ,"r:UTF-8") rescue nil
+      if not file.nil? and file.stat.readable?
+                file.each_line do |line| 
+                        return line.strip.split('::>')[1].chomp.to_s if line.include?('username::>')
+                end
+                return ""
+      end
+      return ""
+    end
+    def get_password
+      return "" if not File.exists?($filepath)
+      file = File.new( $filepath , "r:UTF-8") rescue nil
+      if not file.nil? and file.stat.readable?
+              file.each_line do |line| 
+                      return decrypt(line.strip.split('::>')[1].chomp.to_s) if line.include?('password::>')
+        end
+        return ""
+      end
+      return ""
+    end
+    def get_repeating_seconds
+      return "" if not File.exists?($filepath)
+      file = File.new( $filepath , "r:UTF-8") rescue nil
+      if not file.nil? and file.stat.readable?
+          file.each_line do |line| 
+            return line.strip.split('::>')[1].chomp.to_s if line.include?('repeating_seconds::>')
+          end
+          return ""
+      end
+      return ""
+    end
+#-------------------------------------------#
+
+
+#------------------Cipher-------------------#
+    require 'gibberish'
+    key = "secret-key"
+    $cipher = Gibberish::AES.new(key)
+
+    def encrypt(str)
+        $cipher.enc( str )
+    end
+
+    def decrypt(encrypted_value)
+        $cipher.dec(encrypted_value)
+    end
+#-------------------------------------------#
+
 $filepath = File.join(  ENV['HOME'] ,".userdetails")
 $repeating_seconds = -1
 $showtimer = false
@@ -10,107 +124,22 @@ else
   $showtimer = false
   $repeating_seconds = -1
 end
-require 'open3'
-def cyberoam( username , password , login = true)
-  login_curl = "curl --silent 'https://10.100.56.55:8090/login.xml' -H 'Accept: text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8' -H 'Accept-Encoding: gzip, deflate' -H 'Accept-Language: en-US,en;q=0.5' -H 'Cache-Control: no-cache' -H 'Connection: keep-alive' -H 'Content-Type: application/x-www-form-urlencoded; charset=UTF-8' -H 'Host: 10.100.56.55:8090' -H 'Pragma: no-cache' -H 'Referer: https://10.100.56.55:8090/httpclient.html' --data 'mode=191&username=#{username}&password=#{password}&a=1407179385601&producttype=0' -k --max-time 0.4"
-  logout_curl = "curl --silent 'https://10.100.56.55:8090/logout.xml' -H 'Host: 10.100.56.55:8090' -H 'Accept: text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8' -H 'Accept-Language: en-US,en;q=0.5' -H 'Accept-Encoding: gzip, deflate' -H 'Content-Type: application/x-www-form-urlencoded; charset=UTF-8' -H 'Referer: https://10.100.56.55:8090/httpclient.html' -H 'Connection: keep-alive' -H 'Pragma: no-cache' -H 'Cache-Control: no-cache' --data 'mode=193&username=#{username}&a=1407179351402&producttype=0' -k --max-time 0.4"
-
-  if login
-    loginresponse = `#{login_curl}`                 #works in linux shoes # Not works in windows shoes
-    return loginresponse , (not loginresponse.empty? )
-  else
-    logoutresponse = `#{logout_curl}`                 #Not works
-    return logoutresponse , (not logoutresponse.empty? )
-  end
-end
-def show_notification(title , body)
-    notification = Libnotify.new do |notify|
-    notify.summary    = title
-    notify.body       = body
-    notify.timeout    = 1         
-    notify.urgency    = :critical   
-    notify.append     = true       
-    notify.transient  = false        
-  end
-  notification.show!
-end
-Shoes.setup do
- gem 'nori'
- gem 'nokogiri'
- gem 'libnotify'
-end
-require 'nori'
-require 'nokogiri'
-require 'libnotify'
 
 
-show_notification("Auto CR Login" , "started")
+  def cyberoam( username , password , login = true)
+    login_curl = "curl --silent 'https://10.100.56.55:8090/login.xml' -H 'Accept: text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8' -H 'Accept-Encoding: gzip, deflate' -H 'Accept-Language: en-US,en;q=0.5' -H 'Cache-Control: no-cache' -H 'Connection: keep-alive' -H 'Content-Type: application/x-www-form-urlencoded; charset=UTF-8' -H 'Host: 10.100.56.55:8090' -H 'Pragma: no-cache' -H 'Referer: https://10.100.56.55:8090/httpclient.html' --data 'mode=191&username=#{username}&password=#{password}&a=1407179385601&producttype=0' -k --max-time 0.7"
+    logout_curl = "curl --silent 'https://10.100.56.55:8090/logout.xml' -H 'Host: 10.100.56.55:8090' -H 'Accept: text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8' -H 'Accept-Language: en-US,en;q=0.5' -H 'Accept-Encoding: gzip, deflate' -H 'Content-Type: application/x-www-form-urlencoded; charset=UTF-8' -H 'Referer: https://10.100.56.55:8090/httpclient.html' -H 'Connection: keep-alive' -H 'Pragma: no-cache' -H 'Cache-Control: no-cache' --data 'mode=193&username=#{username}&a=1407179351402&producttype=0' -k --max-time 0.7"
 
-def cache_details(username , password)   
-   repeating_seconds = get_repeating_seconds
-   file = File.new($filepath , 'w') 
-     file.puts "username::>#{username}"
-     file.puts "password::>#{password}"
-     file.puts "repeating_seconds::>#{repeating_seconds}" if not repeating_seconds.strip.empty?
-   file.close
-end
-def save_repeating_time(repeating_seconds)
-   username = get_username
-   password = get_password
-   if( (not username.empty?) and (not password.empty?) )
-       file = File.new($filepath , 'w') 
-         file.puts "username::>#{username}"
-         file.puts "password::>#{password}"
-         file.puts "repeating_seconds::>#{repeating_seconds}"
-       file.close
-       return true
+    if login
+      loginresponse = `#{login_curl}`                 #works in linux shoes # Not works in windows shoes
+      return loginresponse , (not loginresponse.empty? )
     else
-       return false
+      logoutresponse = `#{logout_curl}`                 #Not works
+      return logoutresponse , (not logoutresponse.empty? )
     end
-end
-def unsave_repeating_time
-   username = get_username
-   password = get_password
-   if( (not username.empty?) and (not password.empty?) )
-       file = File.new($filepath , 'w') 
-         file.puts "username::>#{username}"
-         file.puts "password::>#{password}"
-       file.close
-    end
-end
-def get_username
-  return "" if not File.exists?($filepath)
-  file = File.new( $filepath , 'r') rescue nil
-  if not file.nil? and file.stat.readable?
-            file.each_line do |line| 
-                    return line.strip.split('::>')[1].chomp.to_s if line.include?('username::>')
-            end
-            return ""
   end
-  return ""
-end
-def get_password
-  return "" if not File.exists?($filepath)
-  file = File.new( $filepath , 'r') rescue nil
-  if not file.nil? and file.stat.readable?
-          file.each_line do |line| 
-                  return line.strip.split('::>')[1].chomp.to_s if line.include?('password::>')
-    end
-    return ""
-  end
-  return ""
-end
-def get_repeating_seconds
-  return "" if not File.exists?($filepath)
-  file = File.new( $filepath , 'r') rescue nil
-  if not file.nil? and file.stat.readable?
-      file.each_line do |line| 
-        return line.strip.split('::>')[1].chomp.to_s if line.include?('repeating_seconds::>')
-      end
-      return ""
-  end
-  return ""
-end
+
+
 
   def attempt_login(username , password)
     loginresponse , network_available = cyberoam(username , password, true)
@@ -157,13 +186,15 @@ Shoes.app :title => "Auto Cyberoam Login" , :width => 340, :height => 490 , :scr
     self.close
   end
 
-  pkill_installed = `pkill --version` rescue false
-  if not pkill_installed
-    alert("I am Unable to Run pkill command on your system. Please make sure you can run: pkill --version on your terminal.")
-  else
-    `pkill xterm`
+  if not $windows
+    pkill_installed = `pkill --version` rescue false
+    if not pkill_installed
+      alert("I am Unable to Run pkill command on your system. Please make sure you can run: pkill --version on your terminal.")
+    else
+      `pkill xterm`
+    end
   end
-  
+
   stack :margin_left => 40 , :margin_right => 40 , :margin_top => 20 , :margin_bottom => 0 do
     @response = para ""
     stack :margin => 5 do
@@ -253,6 +284,7 @@ Shoes.app :title => "Auto Cyberoam Login" , :width => 340, :height => 490 , :scr
                             $showtimer = true
                             attempt_login( @username.text , @password.text)
                             show_timer(app , 1 , 1)
+                            # @every = every(1){|timer_seconds| @time_elapsed.replace( "Time Elapsed : " + "#{Time.at(timer_seconds).gmtime.strftime('%R:%S')}" , :align => "center")}
                           end
                         end
                       end
@@ -260,6 +292,7 @@ Shoes.app :title => "Auto Cyberoam Login" , :width => 340, :height => 490 , :scr
               else
                       $showtimer = false
                       $repeating_seconds = -1
+                      # @every.stop
                       @para.remove
                       @para = nil
                       @repeating_hours.remove
@@ -292,6 +325,7 @@ Shoes.app :title => "Auto Cyberoam Login" , :width => 340, :height => 490 , :scr
       $showtimer = true
       attempt_login( get_username , get_password)
       show_timer(app , 1 , 1)
+      # @every = every(1){|timer_seconds| @time_elapsed.replace( "Time Elapsed : " + "#{Time.at(timer_seconds).gmtime.strftime('%R:%S')}" , :align => "center")}
       @final_flow.append do
           @link.remove rescue nil
           @link = button "Reset Scheduler Time" , :width => '100%' , :margin => 5 do
@@ -307,6 +341,7 @@ Shoes.app :title => "Auto Cyberoam Login" , :width => 340, :height => 490 , :scr
                     $showtimer = true
                     attempt_login( @username.text , @password.text)
                     show_timer(app , 1 , 1)
+                    # @every = every(1){|timer_seconds| @time_elapsed.replace( "Time Elapsed : " + "#{Time.at(timer_seconds).gmtime.strftime('%R:%S')}" , :align => "center")}
                 end
           end
       end
